Fix issue where dailies don't show up
Had an issue where some of the fractal dailies and wvw didn't show
up correctly. Fixed that issue by filtering out the dailies the site
can't help with earlier on in the same function as API call. That
made the creation of the daily item more predictable.
Definitely room for improvements in the future, but right now it works.

Also fixed issue with daily filtering, so when you click on daily
it should show you correctly search result.