Add recommended fractal, shorten code
Added recommended daily fractals, and made it so you can click
it and it'll get the fractal associated with that level.
Also shortened the code that set results based on click on daily
Before the same code would be repeated based on it was
showing today or tomorrow. Now its one check, and the rest of
the code is only written once.