Fix error in ids array, sent to achievement Api
There was an error in the combining of the arrays with daily ids,
making it so the first and last of the ids within each category wouldn't
be requested from the achievement api. Discovered because the
lowest level recommended fractal never showed up.  Fixed by joining
the different arrays correctly.